% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngPipeline.R
\name{applyPngPipelineOnePage}
\alias{applyPngPipelineOnePage}
\title{Apply the plot preprocessing pipeline}
\usage{
applyPngPipelineOnePage(plotSpec)
}
\arguments{
\item{plotSpec}{A \code{\link[=plotSpec]{plotSpec()}}}
}
\value{
file.path to checked out, extracted and cropped png
}
\description{
Currently, the pipeline has the following steps
}
\details{
\itemize{
\item Check out from git
\item Extract page from pdf as png
\item Crop png to specified part
}
}
