% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExcel.R
\name{availableStyles}
\alias{availableStyles}
\title{Print available styles}
\usage{
availableStyles()
}
\value{
Prints available styles
}
\description{
Print available styles
}
\examples{
availableStyles()
}
\seealso{
Other UI: 
\code{\link{epFiles}()},
\code{\link{parsePlotSpec}()},
\code{\link{parseTextSpec}()},
\code{\link{plotExcel}()},
\code{\link{plotSpec}()},
\code{\link{textSpec}()}
}
\concept{UI}
