% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngPipeline.R
\name{idempotencyNoActionRequired}
\alias{idempotencyNoActionRequired}
\title{Check if anything needs to be done on a plot file}
\usage{
idempotencyNoActionRequired(fileIn, fileOut, commit)
}
\arguments{
\item{fileIn}{Input file of the function where this function is called.}

\item{fileOut}{Output file of the function where this function is called.}

\item{commit}{Commit hash}
}
\value{
TRUE: A sound input output relationship is guaranteed even if we don't redo the step
}
\description{
Check if anything needs to be done on a plot file
}
