% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{onePageMacro}
\alias{onePageMacro}
\title{Print the libre office macro used to put everything on one page with libreoffice headless conversion to pdf}
\usage{
onePageMacro()
}
\value{
Prints the macro
}
\description{
This function is more for documentation purposes, until I find a better way to install it.
}
