% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExcel.R
\name{parseTable}
\alias{parseTable}
\title{Parse the data.table which specifies the texts and plots to be exported}
\usage{
parseTable(d, headerRowStyle)
}
\arguments{
\item{d}{data.table("plot.pdf::spec...", "text::style", ...)}

\item{headerRowStyle}{Default: "center". Style used for header row.}
}
\value{
Augmented \code{d} with parsed information
}
\description{
Parse the data.table which specifies the texts and plots to be exported
}
