% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngPipeline.R
\name{pngPipelineCheckoutToTemp}
\alias{pngPipelineCheckoutToTemp}
\title{Check out a file, potentially from git to a temporary file.}
\usage{
pngPipelineCheckoutToTemp(plotSpec)
}
\arguments{
\item{plotSpec}{A \code{\link[=plotSpec]{plotSpec()}}}
}
\value{
file.path to checked out file. The file name is /tmp/$ABSOLUTEPATH.
}
\description{
Check out a file, potentially from git to a temporary file.
}
