% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngPipeline.R
\name{pngPipelineCrop}
\alias{pngPipelineCrop}
\title{Apply the cropping step of the plot preprocessing pipeline}
\usage{
pngPipelineCrop(plotSpec)
}
\arguments{
\item{plotSpec}{A \code{\link[=plotSpec]{plotSpec()}}}
}
\value{
file.path to cropped png
}
\description{
Apply the cropping step of the plot preprocessing pipeline
}
