% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngPipeline.R
\name{pngPipelineExtractPage}
\alias{pngPipelineExtractPage}
\title{Apply the extraction step of the plot preprocessing pipeline}
\usage{
pngPipelineExtractPage(plotSpec)
}
\arguments{
\item{plotSpec}{A \code{\link[=plotSpec]{plotSpec()}}}
}
\value{
file.path to extracted page as png
}
\description{
Apply the extraction step of the plot preprocessing pipeline
}
