% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExcel.R
\name{populateExcel}
\alias{populateExcel}
\title{Given all parsed information, construct the Excel workbook}
\usage{
populateExcel(dParsed, dwidths, dheights, FLAGaddBorders)
}
\arguments{
\item{dParsed}{Parsed d, data.table}

\item{dwidths}{data.table with column width information}

\item{dheights}{data.table with row height information}

\item{FLAGaddBorders}{Add borders around all cells? Default: FALSE}
}
\value{
An openxlsx workbook \code{wb}
}
\description{
Given all parsed information, construct the Excel workbook
}
