% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolveLockedFilePath}
\alias{resolveLockedFilePath}
\title{Check if file is writable; if not, append a datetime stamp}
\usage{
resolveLockedFilePath(path)
}
\arguments{
\item{path}{File path to check}
}
\value{
Original \code{path} if writable; otherwise, a datetime-stamped path
}
\description{
Check if file is writable; if not, append a datetime stamp
}
