% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPipeline.R
\name{textSpec}
\alias{textSpec}
\title{Collect all text specification in a list}
\usage{
textSpec(text, style)
}
\arguments{
\item{text}{The text you want to insert in the cell}

\item{style}{Numeric or character, specifying the index or the name of a text style. Execute \code{\link[=availableStyles]{availableStyles()}} for available styles.}
}
\value{
List of its input arguments
}
\description{
Collect all text specification in a list
}
\examples{
textSpec(text = "Text 1", style = "rotateUp")
}
\seealso{
Other UI: 
\code{\link{availableStyles}()},
\code{\link{epFiles}()},
\code{\link{parsePlotSpec}()},
\code{\link{parseTextSpec}()},
\code{\link{plotExcel}()},
\code{\link{plotSpec}()}
}
\concept{UI}
