% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateLockedFilename}
\alias{updateLockedFilename}
\title{Check if file can be overwritten and if no add a datetime stamp}
\usage{
updateLockedFilename(filename)
}
\arguments{
\item{filename}{filename to be checked}
}
\value{
\code{filename} if file can be overwritten, and datetime-stamped file if file is locked.
}
\description{
Check if file can be overwritten and if no add a datetime stamp
}
